requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
gunicorn==22.0.0
SQLAlchemy==2.0.35
alembic==1.13.2
pymysql==1.1.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
pydantic==2.9.2
pydantic-settings==2.5.2
jinja2==3.1.4
celery==5.4.0
redis==5.0.8
fpdf2==2.8.1
pandas==2.2.2
